initial downloads of 2020Q1, 2019Q4, and 2018Q4 data

+++ scripts/collect.py
@@ -36,16 +36,20 @@ def fetch_file(url):
 
 def main():
     for filename, url in collect_inventory():
-        stderr.write("\n")
-        stderr.write(f"Downloading: {url}\n")
-
-        content = fetch_file(url)
-
         dest_path = Path(filename)
-        dest_path.parent.mkdir(exist_ok=True, parents=True)
-        dest_path.write_bytes(content)
 
-        stderr.write(f"\tWrote {len(content)} bytes to: {dest_path}\n")
+        stderr.write("\n")
+        if dest_path.is_file() and dest_path.stat().st_size > 1023:
+            stderr.write(f'{dest_path} already exists with {dest_path.stat().st_size} bytes\n')
+        else:
+            stderr.write(f"Downloading: {url}\n")
+
+            content = fetch_file(url)
+
+            dest_path.parent.mkdir(exist_ok=True, parents=True)
+            dest_path.write_bytes(content)
+
+            stderr.write(f"\tWrote {len(content)} bytes to: {dest_path}\n")
 
 
 if __name__ == '__main__':
